backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pydantic-settings
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
apscheduler
python-multipart
email-validator